# premiere-back/src/main/resources/static/api-docs.json
{
  "info": {
    "title": "Premiere API Documentation",
    "version": "0.0.3"
  },
  "openapi": "3.0.1",
  "servers": [
    {
      "url": "http://localhost:8080/",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/auth/reset-password": {
      "put": {
        "summary": "Get back the password that user forgot",
        "tags": [
          "Auth API"
        ],
        "operationId": "resetPassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "username": null,
                  "email": "admin@test.com",
                  "currentPassword": null,
                  "newPassword": "admin1"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "example": {
                  "message": "Check and uncheck exception error message",
                  "i18nPlaceHolder": "PREMIERE.INTERNAL_SERVER_ERROR"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/change-password": {
      "put": {
        "summary": "Change the password of a user",
        "tags": [
          "Auth API"
        ],
        "operationId": "changePassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "username": "admin",
                  "email": null,
                  "currentPassword": "admin",
                  "newPassword": "admin1"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "202": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "example": {
                  "message": "Credit card not found",
                  "i18nPlaceHolder": "CREDIT_CARD.NOT_FOUND"
                }
              }
            }
          },
          "400": {
            "description": "Wrong password",
            "content": {
              "application/json": {
                "example": {
                  "message": "Wrong password",
                  "i18nPlaceHolder": "AUTH.WRONG_PASSWORD"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "example": {
                  "message": "Check and uncheck exception error message",
                  "i18nPlaceHolder": "PREMIERE.INTERNAL_SERVER_ERROR"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/verify-otp": {
      "post": {
        "tags": [
          "Auth API"
        ],
        "operationId": "verifyOTP",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OTPDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "202": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "example": {
                  "message": "Credit card not found",
                  "i18nPlaceHolder": "CREDIT_CARD.NOT_FOUND"
                }
              }
            }
          },
          "409": {
            "description": "Resource already existed",
            "content": {
              "application/json": {
                "example": {
                  "message": "Receiver already existed",
                  "i18nPlaceHolder": "RECEIVER.EXISTED"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "example": {
                  "message": "Check and uncheck exception error message",
                  "i18nPlaceHolder": "PREMIERE.INTERNAL_SERVER_ERROR"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/save-employee": {
      "post": {
        "summary": "Save the employee information",
        "description": "The API can either create or update employee information",
        "tags": [
          "Auth API"
        ],
        "operationId": "saveEmployee",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "username": "employee1",
                  "email": "employee2@test.com",
                  "firstName": "Nhat",
                  "lastName": "Minh",
                  "phone": "0826081111",
                  "gender": "MALE",
                  "panNumber": "1234567899",
                  "address": "Hue",
                  "password": "employee1",
                  "role": "EMPLOYEE"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "example": 1
              }
            }
          },
          "400": {
            "description": "Assign illegal data",
            "content": {
              "application/json": {
                "example": {
                  "message": "Assign illegal data",
                  "i18nPlaceHolder": "AUTH.ROLE.ASSIGN_ILLEGAL_EMPLOYEE_ROLE"
                }
              }
            }
          },
          "409": {
            "description": "Resource already existed",
            "content": {
              "application/json": {
                "example": {
                  "message": "Email already existed",
                  "i18nPlaceHolder": "EMAIL.EXISTED"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "example": {
                  "message": "Check and uncheck exception error message",
                  "i18nPlaceHolder": "PREMIERE.INTERNAL_SERVER_ERROR"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/request-otp": {
      "post": {
        "tags": [
          "Auth API"
        ],
        "operationId": "requestOTP",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OTPDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "202": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "example": {
                  "message": "Credit card not found",
                  "i18nPlaceHolder": "CREDIT_CARD.NOT_FOUND"
                }
              }
            }
          },
          "409": {
            "description": "Resource already existed",
            "content": {
              "application/json": {
                "example": {
                  "message": "Receiver already existed",
                  "i18nPlaceHolder": "RECEIVER.EXISTED"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "example": {
                  "message": "Check and uncheck exception error message",
                  "i18nPlaceHolder": "PREMIERE.INTERNAL_SERVER_ERROR"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/register-customer": {
      "post": {
        "summary": "Register customer",
        "tags": [
          "Auth API"
        ],
        "operationId": "registerCustomer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "username": "minhnhat0212",
                  "email": "lengocminhnhatntp@gmail.com",
                  "firstName": "Minh Nhat",
                  "lastName": "Le",
                  "phone": "0123456722",
                  "gender": "MALE",
                  "panNumber": "19191196549",
                  "address": "Ho Chi Minh",
                  "password": "minhnhat0212",
                  "role": "CUSTOMER"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "202": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "example": {
                  "message": "User not found",
                  "i18nPlaceHolder": "USER.NOT_FOUND"
                }
              }
            }
          },
          "400": {
            "description": "Assign illegal data",
            "content": {
              "application/json": {
                "example": {
                  "message": "Assign illegal data",
                  "i18nPlaceHolder": "AUTH.ROLE.ASSIGN_ILLEGAL_EMPLOYEE_ROLE"
                }
              }
            }
          },
          "409": {
            "description": "Resource already existed",
            "content": {
              "application/json": {
                "example": {
                  "message": "Phone number already existed",
                  "i18nPlaceHolder": "PHONE.EXISTED"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "example": {
                  "message": "Check and uncheck exception error message",
                  "i18nPlaceHolder": "PREMIERE.INTERNAL_SERVER_ERROR"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/token/user": {
      "get": {
        "summary": "Get user info by token",
        "tags": [
          "Auth API"
        ],
        "operationId": "getUserByToken",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "202": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "example": {
                  "message": "Credit card not found",
                  "i18nPlaceHolder": "CREDIT_CARD.NOT_FOUND"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "example": {
                  "message": "Check and uncheck exception error message",
                  "i18nPlaceHolder": "PREMIERE.INTERNAL_SERVER_ERROR"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/get-employees": {
      "get": {
        "tags": [
          "Auth API"
        ],
        "summary": "Get all employees",
        "operationId": "getAllEmployees",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "example": [
                    {
                      "id": 2,
                      "version": 0,
                      "createdAt": "2023-01-12 07:16:00",
                      "updatedAt": "2023-01-12 07:16:00",
                      "username": "customer2",
                      "email": "customer2@gmail.com",
                      "firstName": "Giap",
                      "lastName": "Hoang Huu",
                      "phone": "0987654322",
                      "gender": "FEMALE",
                      "panNumber": "pan_number",
                      "address": "address",
                      "password": null,
                      "role": "CUSTOMER",
                      "enabled": true,
                      "cardEnabled": true
                    },
                    {
                      "id": 5,
                      "version": 0,
                      "createdAt": "2023-01-12 07:16:00",
                      "updatedAt": "2023-01-12 07:16:00",
                      "username": "customer5",
                      "email": "customer5@gmail.com",
                      "firstName": "Hai",
                      "lastName": "Nguyen Van",
                      "phone": "0987654325",
                      "gender": "MALE",
                      "panNumber": "pan_number",
                      "address": "address",
                      "password": null,
                      "role": "CUSTOMER",
                      "enabled": true,
                      "cardEnabled": true
                    }
                  ],
                  "items": {
                    "$ref": "#/components/schemas/ReceiverDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "example": {
                  "message": "Check and uncheck exception error message",
                  "i18nPlaceHolder": "PREMIERE.INTERNAL_SERVER_ERROR"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ReceiverDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "cardNumber": {
            "type": "string"
          },
          "nickname": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "bankName": {
            "type": "string"
          }
        }
      },
...
      "OTPDto": {
        "type": "object",
        "properties": {
          "otp": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "CUSTOMER",
              "EMPLOYEE",
              "PREMIERE_ADMIN"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}